refactor: remoção de componentes padrões e conteúdos do projeto que não faziam sentido no momento para criação das FEATURES
